Crate lipmaa_link

A function to calculate lipmaa sequence numbers.

From the bamboo spec: "The lipmaalinks are chosen such that for any pair of entries there is a path from the newer to the older one of a length logarithmic in their distance."

use lipmaa_link::lipmaa;

let result = lipmaa(13);
assert_eq!(result, 4);

Functions

lipmaa
Calculates the lipmaa link number given the current sequence number.